+++ AndroidStudio/assignments/fundamentals/6th/ListViewAssignment/src/main/res/layout/activity_book.xml
@@ -6,7 +6,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".MainActivity" >
+    tools:context=".BookActivity" >
 
     <TextView
         android:id="@+id/BookTitle"
