+++ AndroidStudio/assignments/fundamentals/6th/ListViewAssignment/src/main/res/layout/activity_list_view.xml
@@ -1,0 +1,12 @@
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".ListViewActivity" >
+
+    <ListView
+        android:id="@+id/BookList"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
+
+</RelativeLayout>
