+++ AndroidStudio/assignments/fundamentals/6th/ListViewAssignment/src/main/res/layout/activity_main.xml
@@ -1,12 +1,14 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity" >
+    android:orientation="vertical"
+    tools:context=".MainActivity">
 
-    <ListView
-        android:id="@+id/BookList"
+    <Button
+        android:id="@+id/list_view1_button"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-</RelativeLayout>
+        android:layout_height="wrap_content"
+        android:text="@string/list_view_activity_name" />
+</LinearLayout>
